requirements.txt
Flask>=2.0
requests>=2.20
pytest>=7.0
Flask-SQLAlchemy>=3.0
Flask-Session>=0.5
redis>=4.0
python-dotenv>=1.0
Flask-Migrate>=4.0
werkzeug>=2.1
Flask-Login>=0.6
Pillow>=9.0
flake8>=6.0
scikit-learn>=1.3
Flask-Cors>=4.0
pyngrok>=7.0